+++ src/ghrepo/__init__.py
@@ -172,7 +172,7 @@ def get_local_repo(dirpath: Optional[AnyPath] = None, remote: str = "origin") ->
     containing the directory ``dirpath`` (default: the current directory) by
     parsing the URL for the specified remote.  Raises a
     `subprocess.CalledProcessError` if the given path is not in a GitHub
-    repository.
+    repository or the given remote does not exist.
     """
     r = subprocess.run(
         ["git", "remote", "get-url", "--", remote],
@@ -182,6 +182,28 @@ def get_local_repo(dirpath: Optional[AnyPath] = None, remote: str = "origin") ->
         check=True,
     )
     return GHRepo.parse_url(r.stdout.strip())
+
+
+def get_branch_upstream(branch: str, dirpath: Optional[AnyPath] = None) -> GHRepo:
+    """
+    .. versionadded:: 0.5.0
+
+    Determine the GitHub repository for the upstream remote of the given branch
+    in the Git repository located at or containing the directory ``dirpath``
+    (default: the current directory).
+
+    Raises a `subprocess.CalledProcessError` if the given path is not in a
+    GitHub repository or the given branch does not have an upstream remote
+    configured.
+    """
+    upstream = subprocess.run(
+        ["git", "config", "--get", "--", f"branch.{branch}.remote"],
+        cwd=dirpath,
+        stdout=subprocess.PIPE,
+        universal_newlines=True,
+        check=True,
+    ).stdout.strip()
+    return get_local_repo(dirpath, remote=upstream)
 
 
 def get_current_branch(dirpath: Optional[AnyPath] = None) -> str: